backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-cloud-vision==3.4.4
google-cloud-speech==2.21.0
google-cloud-aiplatform==1.35.0
google-cloud-translate==3.12.1
Werkzeug==2.3.7
Pillow==10.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
